Added user role requirement checking. Refactoring.
