+++ app/views/views.py
@@ -1,8 +1,10 @@
-from flask import render_template, flash, redirect, g, url_for, session, request
-from flask.ext.login import current_user, login_required, login_user, logout_user
+from flask import render_template, flash, redirect, g, url_for, session
+from flask.ext.login import current_user, login_required, logout_user
+
 from app import app, db, lm
 from app.forms.login import LoginForm
 from app.models import User
+from app.views.util.login import tryLogin, requires_roles
 
 
 @app.route('/')
@@ -13,19 +15,6 @@ def index():
     return render_template("index.html",
                            title='Home',
                            user=user)
-
-
-def tryLogin(user, password):
-    if user and user.password == password:
-        remember_me = False
-        if 'remember_me' in session:
-            remember_me = session['remember_me']
-            session.pop('remember_me', None)
-        login_user(user, remember=remember_me)
-        return redirect(request.args.get('next') or url_for('index'))
-    else:
-        flash("Invalid login. Please try again.")
-        return redirect(url_for('index'))
 
 
 @app.route('/login', methods=['GET', 'POST'])
@@ -66,6 +55,13 @@ def test_login():
     return 'You are logged in as ' + g.user.user_name
 
 
+@app.route('/test_admin')
+@login_required
+@requires_roles('a')
+def test_admin():
+    return g.user.user_name + " is an admin!"
+
+
 @lm.user_loader
 def load_user(id):
     return User.query.get(id)
@@ -73,3 +69,5 @@ def load_user(id):
 @app.before_request
 def before_request():
     g.user = current_user
+
+
